Renamed files to more consistent names
Renamed ci_lazy_methods to ci_lazy_methods_needed and
gather_methods_lazily to ci_lazy_methods
