+++ src/java_bytecode/java_bytecode_convert_method.h
@@ -18,7 +18,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include "java_string_library_preprocess.h"
 
 #include "java_bytecode_parse_tree.h"
-#include "ci_lazy_methods.h"
+#include <java_bytecode/ci_lazy_methods_needed.h>
 
 class class_hierarchyt;
 
@@ -28,7 +28,7 @@ void java_bytecode_convert_method(
   symbol_tablet &symbol_table,
   message_handlert &message_handler,
   size_t max_array_length,
-  safe_pointer<ci_lazy_methodst> lazy_methods,
+  safe_pointer<ci_lazy_methods_neededt> lazy_methods,
   java_string_library_preprocesst &string_preprocess);
 
 inline void java_bytecode_convert_method(
@@ -45,7 +45,7 @@ inline void java_bytecode_convert_method(
     symbol_table,
     message_handler,
     max_array_length,
-    safe_pointer<ci_lazy_methodst>::create_null(),
+    safe_pointer<ci_lazy_methods_neededt>::create_null(),
     string_preprocess);
 }
